feat(b20): V0.2.0 ISA Phase 6 - control-flow operand narrowing plus native dedup
Phase 6 of the V0.2.0 ISA implementation (B20). The six block-
structured control-flow opcodes carry `u16` jump targets instead
of `u32`, capping any single chunk at 65,535 ops. A new
`compile_with_warnings` entry point exposes the soft-warning
threshold at 80% of the cap. The Phase 5 follow-on for concern
#2 (multiple native registrations of the same name) is included
in the same commit: `register_*` methods now deduplicate.

Opcodes

- `Op::If(u16)`, `Op::Else(u16)`, `Op::Loop(u16)`,
  `Op::EndLoop(u16)`, `Op::Break(u16)`, `Op::BreakIf(u16)`
  replace the previous `u32`-targeted variants. The rkyv
  ArchivedOp tracks the change automatically through the
  derive.
- Approximately 18 `as u32` arithmetic sites in `src/compiler.rs`
  rewrote to `as u16` (loop counter increment, EndLoop back-
  edge, after-EndLoop, after-Loop, branch patches). The cast
  in `FuncCompiler::patch_jump` likewise narrows; the post-
  emit hard-cap check guarantees the cast never truncates an
  admissible chunk.
- VM dispatch arms continue to widen `u16 as usize` unchanged.

Compile-time enforcement

- New `pub struct CompileWarning { message: String, chunk_name:
  String }` in `src/compiler.rs`.
- New constants `CHUNK_SIZE_HARD_LIMIT = u16::MAX as usize`
  (65,535) and `CHUNK_SIZE_SOFT_WARN_THRESHOLD = 52,428` (80%
  of the hard limit).
- New `pub fn compile_with_warnings(program, target) ->
  Result<(Module, Vec<CompileWarning>), CompileError>` runs the
  full compile pipeline and returns the warnings vector
  alongside the module.
- `compile_with_target` is now a thin wrapper that delegates to
  `compile_with_warnings` and discards the warnings. The 30+
  existing call sites preserve their signature.
- Hard cap: any chunk whose op count exceeds
  `CHUNK_SIZE_HARD_LIMIT` produces a `CompileError` with the
  offending function's source span; the diagnostic prompts
  decomposition into helpers.
- Soft warning: any chunk crossing
  `CHUNK_SIZE_SOFT_WARN_THRESHOLD` produces one
  `CompileWarning` in the returned vector.

Concern #2: dedup native registration

- The eight `register_*` methods on `GenericVm`
  (`register_native`, `register_native_closure`,
  `register_native_with_ctx`, `register_native_with_ctx_-
  closure`, `register_fn`, `register_fn_fallible`,
  `register_verified_native`, `register_external_native`) now
  call `self.natives.retain(|e| e.name != name)` before pushing
  the new entry. A re-registration of the same name replaces
  the prior entry rather than appending.
- Combined with the cache invalidation introduced in the Phase
  5 follow-on, this makes re-registration semantically
  predictable: the latest registration is the one the dispatch
  observes, and the load-time classification check re-runs.

Tests

- `compiler::tests::chunk_size_thresholds_are_consistent` pins
  the relationship between the constants.
- `compiler::tests::small_chunk_produces_no_warnings` covers
  the empty-warnings path through
  `compile_with_warnings`.
- `vm::tests::duplicate_native_registration_replaces_prior_entry`
  exercises the dedup happy path: two `register_native` calls
  with the same name leave only the second function visible.
- `vm::tests::duplicate_native_registration_swaps_classification`
  exercises classification swap on re-registration: a verified
  registration replaced by an external registration produces no
  mismatch even though the script imports `use external`.

Verification

- 759 keleusma lib tests pass (up from 755 because the four new
  tests were added).
- 53 rogue-script integration tests pass.
- 17 marshall tests pass.
- Workspace tests all green.
- cargo clippy --tests --all-targets -- -D warnings clean.
- cargo build --examples clean.
- STM32N6570-DK full pipeline release build (cargo build
  --release --bin three-task-n6 --target
  thumbv8m.main-none-eabihf --no-default-features --features
  stm32n6570dk-platform) clean.

+++ src/bytecode.rs
@@ -448,20 +448,28 @@ pub enum Op {
 
     // -- Block-structured control flow --
     /// Pop bool; if false, skip to target (matching Else or EndIf).
-    If(u32),
-    /// Skip to target (matching EndIf). Reached when then-block falls through.
-    Else(u32),
+    /// Target is an op index within the current chunk; chunks are
+    /// capped at `u16::MAX` ops by the compiler.
+    If(u16),
+    /// Skip to target (matching EndIf). Reached when then-block
+    /// falls through. Target is an op index within the current
+    /// chunk.
+    Else(u16),
     /// Block delimiter for If/Else. No-op at runtime.
     EndIf,
 
-    /// Begin loop block. Target is past EndLoop (used by Break/BreakIf).
-    Loop(u32),
-    /// Back-edge to instruction after matching Loop.
-    EndLoop(u32),
-    /// Unconditional forward jump past enclosing EndLoop.
-    Break(u32),
+    /// Begin loop block. Target is past EndLoop (used by Break and
+    /// BreakIf). Target is an op index within the current chunk.
+    Loop(u16),
+    /// Back-edge to instruction after matching Loop. Target is an
+    /// op index within the current chunk.
+    EndLoop(u16),
+    /// Unconditional forward jump past enclosing EndLoop. Target is
+    /// an op index within the current chunk.
+    Break(u16),
     /// Pop bool; if true, forward jump past enclosing EndLoop.
-    BreakIf(u32),
+    /// Target is an op index within the current chunk.
+    BreakIf(u16),
 
     // -- Streaming --
     /// Stream block entry marker. No-op at runtime.
